added in the scroll and table selection functionality
need to clean this up and move components to new files. kinda ugly as is right now.
also need to think up a better color scheme. couldnt figure out how to make
this work with a StatefulWidget where state is carried throughout. eventually
might need to mess around with that and add it back in but for now this method works.